+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## 0.0.2 - 2018-04-18
+### Changed
+- Added future blockhash as an additional source of participant selection entropy
+- Added an entropy parameter for calculating the final random from the pool of participants
+
 ## 0.0.1 - 2018-05-07
 ### Added
 - Starting to track changes in this log as project moves into beta status